// src/contracts/generated/csi10/BootstrapRequestV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Secret32 } from "./Secret32";

export type BootstrapRequestV1 = { "kind": "returnSameTab", schemaVersion: number, returnReceipt: string, expectedTabTransactionId: Secret32, } | { "kind": "returnFreshTab", schemaVersion: number, returnReceipt: string, } | { "kind": "ordinary", schemaVersion: number, accountSlotId: string, };